Move backend seed files to dedicated folder and update tests accordingly. Tidied up frontend a little and updated README with current state of the the project
