--- frontend/src/components/Header/Layout.jsx
@@ -1,2 +1,0 @@
-// Backward compatibility wrapper - delegates to TypeScript implementation
-export { default } from './Layout.tsx';
